pyproject.toml
[project]
name = "datafusion-nyc-2025-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arro3-core>=0.6.2",
    "geoarrow-rust-core>=0.5.2",
    "geodatafusion>=0.1.1",
    "jupyterlab>=4.4.7",
    "lonboard>=0.12.0",
    "marimo>=0.15.4",
    "matplotlib>=3.10.6",
    "palettable>=3.3.3",
    "pyarrow>=21.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
